function-negative-sentiment/requirements.txt
gunicorn
flask
google-cloud-language
google-cloud-secret-manager
requests

